+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- CurrentWidget::get_block() to resolve the block a block widget holds
+- CurrentWidget::get_type() to tell a block widget from a legacy one
+
 ## [1.2.0] - 2026-07-09
 
 ### Added